refactor: move validateArgs and components to separate files
