+++ cmd/components.go
@@ -1,0 +1,7 @@
+package cmd
+
+const (
+	pipeline  = "pipeline"
+	dashboard = "dashboard"
+	triggers  = "triggers"
+)
