+++ cmd/install.go
@@ -14,12 +14,6 @@ var (
 	dashboardVersion string
 )
 
-const (
-	pipeline  = "pipeline"
-	dashboard = "dashboard"
-	triggers  = "triggers"
-)
-
 var installCmd = &cobra.Command{
 	Use:   "install",
 	Short: "Install various components of Tekton",
@@ -100,25 +94,6 @@ func install(args []string) error {
 		fmt.Printf("Component %s has been installed successfully\n", arg)
 	}
 
-	return nil
-}
-
-func validateArgs(args []string) error {
-	validComponents := make(map[string]bool)
-	validComponents[pipeline] = true
-	validComponents[triggers] = true
-	validComponents[dashboard] = true
-	validComponents["all"] = true
-
-	for _, arg := range args {
-		if !validComponents[arg] {
-			return fmt.Errorf("invalid argument provided to command: %s", arg)
-		}
-
-		if arg == "all" && args[0] != "all" {
-			return fmt.Errorf("all should be only argument provided when used")
-		}
-	}
 	return nil
 }
